+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] 2026-02-25
+### Add
+- Aggiunge i pulsanti 'Visualizza Volume' e 'Modifica Volume' alla sezione 'Gestione Capitoli'
+### Changed
+- Sostituisce 'Numero Capitolo' con 'Indice Capitolo' rendendolo un campo testuale.
+
 ## [1.3.1] 2026-02-14
 ### Add
 - Aggiunge il pulsante 'Gestisci Volume' dalla Modifica Volume
@@ -72,7 +78,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Cleanup controllato in deactivation/uninstall.
 
 
-[Unreleased]: https://github.com/erremauro/cz-volume/compare/v1.3.1...HEAD
+[Unreleased]: https://github.com/erremauro/cz-volume/compare/v1.4.0...HEAD
+[1.4.0]: https://github.com/erremauro/cz-volume/releases/tag/v1.4.0
 [1.3.1]: https://github.com/erremauro/cz-volume/releases/tag/v1.3.1
 [1.3.0]: https://github.com/erremauro/cz-volume/releases/tag/v1.3.0
 [1.2.1]: https://github.com/erremauro/cz-volume/releases/tag/v1.2.1